lib: Update hash.[ch] from latest gnulib
The main difference is the relicense to LGPL
